Small refactorizations and fixes: - Fixed line ending of groups.lst - Fixed GUI behaviour in race_select.cpp - Removed redundant small menu title, use large one instead - Code cleanups
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@550 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ src/gui.h
@@ -210,7 +210,7 @@ void DrawCursor();
 
 // --------------------------------------------------------------------
 
-int AutoYPosN(double percent);
-TArea AutoAreaN(double top_perc, double bott_perc, unsigned int w);
+int AutoYPosN(int percent);
+TArea AutoAreaN(int top_perc, int bott_perc, unsigned int w);
 
 #endif
